+++ system/src/Grav/Common/Page/Interfaces/PageInterface.php
@@ -1,0 +1,9 @@
+<?php
+namespace Grav\Common\Page\Interfaces;
+
+/**
+ * Class implements page interface.
+ */
+interface PageInterface
+{
+}